poststart-hook, hotdog-cc-hook: ensure hotpatch files are injected appropriately
Now the prestart hook will fail if any of the directories in the mount's
target path isn't a directory.

A sentry file was added inside the containers's bundle to indicate that the
prestart hook successfully set up the container to execute the hot patch.

The poststart hook won't attempt to execute any binary in the expected
bundle directory if the sentry file created by the prestart hook
doesn't exist.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>
Reviewed-by: Ben Cressey <bcressey@amazon.com>
Reviewed-by: Samuel Karp <skarp@amazon.com>